requirements.txt
django
djangorestframework
djangorestframework-simplejwt
djoser
pillow
django-imagekit